major rock till cleanup into baseclass
