+++ terrainbento/base_class/two_lithology_erosion_model.py
@@ -6,6 +6,7 @@ Base class for common functions of terrainbentostochastic erosion models.
 The **TwoLithologyErosionModel** is a base class that contains all of the
 functionality shared by the terrainbento models that have two lithologies.
 """
+import numpy as np
 
 from landlab.io import read_esri_ascii
 from terrainbento.base_class import ErosionModel
@@ -57,15 +58,30 @@ class TwoLithologyErosionModel(ErosionModel):
             BoundaryHandlers=BoundaryHandlers,
             OutputWriters=OutputWriters,
         )
+
+        # Get all common parameters
         self.contact_width = (self._length_factor) * self.params[
             "contact_zone__width"
         ]  # has units length
 
         self.m = self.params["m_sp"]
         self.n = self.params["n_sp"]
+
         self.regolith_transport_parameter = (
             self._length_factor ** 2.
         ) * self.get_parameter_from_exponent("regolith_transport_parameter")
+
+        self.K_rock = self.get_parameter_from_exponent("water_erodability~lower") * (
+            self._length_factor ** (1. - (2. * self.m))
+        )
+
+        self.K_till = self.get_parameter_from_exponent("water_erodability~upper") * (
+            self._length_factor ** (1. - (2. * self.m))
+        )
+
+        # Set the erodability values, these need to be double stated because a PrecipChanger may adjust them
+        self.rock_erody = self.K_rock
+        self.till_erody = self.K_till
 
     def _setup_contact_elevation(self):
         file_name = self.params["lithology_contact_elevation__file_name"]
@@ -76,12 +92,91 @@ class TwoLithologyErosionModel(ErosionModel):
         )
         self.rock_till_contact = self.grid.at_node["rock_till_contact__elevation"]
 
+    def _setup_rock_and_till(self):
+        """Set up fields to handle for two layers with different erodability."""
+        # Get a reference to the rock-till field
+        self._setup_contact_elevation()
+
+        # Create field for erodability
+        self.erody = self.grid.add_zeros("node", "substrate__erodability")
+
+        # Create array for erodability weighting function
+        self.erody_wt = np.zeros(self.grid.number_of_nodes)
+
+        # Set values correctly
+        self._update_erodability_field()
+
+    def _setup_rock_and_till_with_threshold(self):
+        """Set up fields to handle for two layers with different erodability."""
+        # Get a reference to the rock-till field\
+        self._setup_contact_elevation()
+
+        # Create field for erodability
+        self.erody = self.grid.add_zeros("node", "substrate__erodability")
+
+        # Create field for threshold values
+        self.threshold = self.grid.add_zeros("node", "water_erosion_rule__threshold")
+
+        # Create array for erodability weighting function
+        self.erody_wt = np.zeros(self.grid.number_of_nodes)
+
+        # set values correctly
+        self._update_erodability_and_threshold_fields()
+
     def _update_erodywt(self):
         # Update the erodability weighting function (this is "F")
-        self.erody_wt[self.data_nodes] = 1.0 / (
-            1.0
-            + np.exp(
-                -(self.z[self.data_nodes] - self.rock_till_contact[self.data_nodes])
-                / self.contact_width
+        core = self.grid.core_nodes
+        if self.contact_width > 0.0:
+            self.erody_wt[core] = 1.0 / (
+                1.0
+                + np.exp(
+                    -(self.z[core] - self.rock_till_contact[core]) / self.contact_width
+                )
             )
+        else:
+            self.erody_wt[core] = 0.0
+            self.erody_wt[np.where(self.z > self.rock_till_contact)[0]] = 1.0
+
+    def _update_Ks_with_precip(self):
+        # (if we're varying K through time, update that first)
+        if "PrecipChanger" in self.boundary_handler:
+            erode_factor = self.boundary_handler[
+                "PrecipChanger"
+            ].get_erodability_adjustment_factor()
+            self.till_erody = self.K_till * erode_factor
+            self.rock_erody = self.K_rock * erode_factor
+
+    def _update_erodability_field(self):
+        """Update erodability at each node.
+
+        The erodability at each node is a smooth function between the rock and
+        till erodabilities and is based on the contact zone width and the
+        elevation of the surface relative to contact elevation.
+        """
+        self._update_erodywt()
+        self._update_Ks_with_precip()
+
+        # Calculate the effective erodibilities using weighted averaging
+        self.erody[:] = (
+            self.erody_wt * self.till_erody + (1.0 - self.erody_wt) * self.rock_erody
+        )
+
+    def _update_erodability_and_threshold_fields(self):
+        """Update erodability at each node.
+
+        The erodability at each node is a smooth function between the rock and
+        till erodabilities and is based on the contact zone width and the
+        elevation of the surface relative to contact elevation.
+        """
+        self._update_erodywt()
+        self._update_Ks_with_precip()
+
+        # Calculate the effective erodibilities using weighted averaging
+        self.erody[:] = (
+            self.erody_wt * self.till_erody + (1.0 - self.erody_wt) * self.rock_erody
+        )
+
+        # Calculate the effective thresholds using weighted averaging
+        self.threshold[:] = (
+            self.erody_wt * self.till_thresh + (1.0 - self.erody_wt) * self.rock_thresh
         )
