+++ terrainbento/derived_models/model_842_basicChRtTh.py
@@ -196,29 +196,16 @@ class BasicChRtTh(TwoLithologyErosionModel):
             OutputWriters=OutputWriters,
         )
 
-        self.K_rock_sp = self.get_parameter_from_exponent("water_erodability~lower") * (
-            self._length_factor ** (1. - (2. * self.m))
-        )
-        self.K_till_sp = self.get_parameter_from_exponent("water_erodability~upper") * (
-            self._length_factor ** (1. - (2. * self.m))
-        )
-        rock_erosion__threshold = self.get_parameter_from_exponent(
+        # Save the threshold values for rock and till
+        self.rock_thresh = self.get_parameter_from_exponent(
             "water_erosion_rule~lower__threshold"
         )
-        till_erosion__threshold = self.get_parameter_from_exponent(
+        self.till_thresh = self.get_parameter_from_exponent(
             "water_erosion_rule~upper__threshold"
         )
 
-        # Set the erodability values, these need to be double stated because a PrecipChanger may adjust them
-        self.rock_erody = self.K_rock_sp
-        self.till_erody = self.K_till_sp
-
-        # Save the threshold values for rock and till
-        self.rock_thresh = rock_erosion__threshold
-        self.till_thresh = till_erosion__threshold
-
         # Set up rock-till boundary and associated grid fields.
-        self._setup_rock_and_till()
+        self._setup_rock_and_till_with_threshold()
 
         # Instantiate a StreamPowerSmoothThresholdEroder component
         self.eroder = StreamPowerSmoothThresholdEroder(
@@ -235,48 +222,6 @@ class BasicChRtTh(TwoLithologyErosionModel):
             linear_diffusivity=self.regolith_transport_parameter,
             slope_crit=self.params["critical_slope"],
             nterms=7,
-        )
-
-    def _setup_rock_and_till(self):
-        """Set up fields to handle for two layers with different erodability."""
-        # Get a reference to the rock-till field\
-        self._setup_contact_elevation()
-
-        # Create field for erodability
-        self.erody = self.grid.add_zeros("node", "substrate__erodability")
-
-        # Create field for threshold values
-        self.threshold = self.grid.add_zeros("node", "water_erosion_rule__threshold")
-
-        # Create array for erodability weighting function
-        self.erody_wt = np.zeros(self.grid.number_of_nodes)
-
-    def _update_erodability_and_threshold_fields(self):
-        """Update erodability at each node.
-
-        The erodability at each node is a smooth function between the rock and
-        till erodabilities and is based on the contact zone width and the
-        elevation of the surface relative to contact elevation.
-        """
-        # Update the erodability weighting function (this is "F")
-        self._update_erodywt()
-
-        # (if we're varying K through time, update that first)
-        if "PrecipChanger" in self.boundary_handler:
-            erode_factor = self.boundary_handler[
-                "PrecipChanger"
-            ].get_erodability_adjustment_factor()
-            self.till_erody = self.K_till_sp * erode_factor
-            self.rock_erody = self.K_rock_sp * erode_factor
-
-        # Calculate the effective erodibilities using weighted averaging
-        self.erody[:] = (
-            self.erody_wt * self.till_erody + (1.0 - self.erody_wt) * self.rock_erody
-        )
-
-        # Calculate the effective thresholds using weighted averaging
-        self.threshold[:] = (
-            self.erody_wt * self.till_thresh + (1.0 - self.erody_wt) * self.rock_thresh
         )
 
     def run_one_step(self, dt):
