+++ terrainbento/derived_models/model_C00_basicRtSa.py
@@ -214,16 +214,16 @@ class BasicRtSa(TwoLithologyErosionModel):
             BoundaryHandlers=BoundaryHandlers,
             OutputWriters=OutputWriters,
         )
-        self.K_rock_sp = self.get_parameter_from_exponent("water_erodability~lower") * (
+        self.K_rock = self.get_parameter_from_exponent("water_erodability~lower") * (
             self._length_factor ** (1. - (2. * self.m))
         )
-        self.K_till_sp = self.get_parameter_from_exponent("water_erodability~upper") * (
+        self.K_till = self.get_parameter_from_exponent("water_erodability~upper") * (
             self._length_factor ** (1. - (2. * self.m))
         )
 
         # Set the erodability values, these need to be double stated because a PrecipChanger may adjust them
-        self.rock_erody = self.K_rock_sp
-        self.till_erody = self.K_till_sp
+        self.rock_erody = self.K_rock
+        self.till_erody = self.K_till
 
         # Set up rock-till boundary and associated grid fields.
         self._setup_rock_and_till()
@@ -268,39 +268,6 @@ class BasicRtSa(TwoLithologyErosionModel):
             self.grid,
             soil_production__maximum_rate=max_soil_production_rate,
             soil_production__decay_depth=soil_production_decay_depth,
-        )
-
-    def _setup_rock_and_till(self):
-        """Set up fields to handle for two layers with different erodability."""
-        # Get a reference to the rock-till field\
-        self._setup_contact_elevation()
-
-        # Create field for erodability
-        self.erody = self.grid.add_zeros("node", "substrate__erodability")
-
-        # Create array for erodability weighting function
-        self.erody_wt = np.zeros(self.grid.number_of_nodes)
-
-    def _update_erodability_field(self):
-        """Update erodability at each node.
-
-        The erodability at each node is a smooth function between the rock and
-        till erodabilities and is based on the contact zone width and the
-        elevation of the surface relative to contact elevation.
-        """
-        self._update_erodywt()
-
-        # (if we're varying K through time, update that first)
-        if "PrecipChanger" in self.boundary_handler:
-            erode_factor = self.boundary_handler[
-                "PrecipChanger"
-            ].get_erodability_adjustment_factor()
-            self.till_erody = self.K_till_sp * erode_factor
-            self.rock_erody = self.K_rock_sp * erode_factor
-
-        # Calculate the effective erodibilities using weighted averaging
-        self.erody[:] = (
-            self.erody_wt * self.till_erody + (1.0 - self.erody_wt) * self.rock_erody
         )
 
     def run_one_step(self, dt):
